global IDF bigint total documents storing
git-svn-id: svn://svn.sphinxsearch.com/sphinx/trunk@3581 406a0c4d-033a-0410-8de8-e80135713968

+++ src/sphinx.h
@@ -3070,7 +3070,7 @@ protected:
 
 	CSphVector<IDFWord_t>	m_dWords;
 	CSphVector<int>			m_dHash;
-	int						m_iTotalDocuments;
+	int64_t					m_iTotalDocuments;
 };
 
 /////////////////////////////////////////////////////////////////////////////
